整合Redis整机替换(master/slave/proxy) #133
