+++ dbm-ui/backend/flow/engine/bamboo/scene/redis/atom_jobs/__init__.py
@@ -9,6 +9,8 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
+from .redis_cluster_master_rep import RedisClusterMasterReplaceJob
+from .redis_cluster_slave_rep import RedisClusterSlaveReplaceJob
 from .redis_dbmon import RedisDbmonAtomJob
 from .redis_install import RedisBatchInstallAtomJob
 from .redis_makesync import RedisMakeSyncAtomJob
